Implementation of the queries, mutations, resolvers and repositories Changed graphql library

+++ models/User.go
@@ -2,7 +2,7 @@ package models
 
 //users
 type User struct {
-	Id        int64      `json:"id"`
+	Id        int        `json:"id"`
 	Name      string     `json:"name"`
 	Email     string     `json:"email"`
 	Posts     []Post     `json:"posts"`
@@ -10,5 +10,5 @@ type User struct {
 }
 
 type Follower struct {
-	UserId int64 `json:"user_id"`
+	UserId int `json:"user_id"`
 }
